BUG: Accounts.setPassword() fails due to aws-sdk
* if you install the npm `aws-sdk` but do not import it, no problem
* if you import `aws-sdk` but do not actually use it somewhere, BUG
* if you import `aws-sdk` and use it (not shown) AWS works, but the Accounts BUG
* if you require `aws-skd` (ES5) BUG

It was a few hours to whittle down this bug to this cause.

I do not know what inside of `aws-sdk` would be causing this problem...

It might be only broken with recent versions of `aws-sdk`
(unsure, I'll be testing a few soon, and will update if I find any working)

It is likely an npm issue (?) but if it's happening for AWS, it's important,
and perhaps indicative of a broader issue.